fix: questions.buyer_id Integer -> BigInteger (migration 0029)
Detectado pelo runtime-watcher no ciclo 457 apos corrigir o
ml_question_id. ML user IDs tambem excedem int32 (ex 2490612425).

Mesmo padrao do bug 0028. Varredura nos outros models nao encontrou
mais ocorrencias do padrao Integer() em campos *_id externos.

+++ backend/app/perguntas/models.py
@@ -48,7 +48,7 @@ class Question(Base):
         comment="UNANSWERED | ANSWERED | CLOSED_UNANSWERED | UNDER_REVIEW",
     )
     buyer_id: Mapped[int | None] = mapped_column(
-        Integer(), nullable=True, comment="ML user ID do comprador"
+        BigInteger(), nullable=True, comment="ML user ID do comprador"
     )
     buyer_nickname: Mapped[str | None] = mapped_column(
         String(100), nullable=True, comment="Nickname do comprador no ML"
